requirements.txt
streamlit
pandas
openai
playwright
python-dotenv
google-generativeai
sqlite3
asyncio
